feat(backend): remove unnecessary db.serialize and add debug
